Significantly speed up all schema changes
Cassandra drivers tend to wait for the schema to settle across all nodes
whenever we execute DDL statements like CREATE/ALTER/DROP. On a single
node cluster we know that the schema settled by the time the query
finished executing. We can thus disable some of the waiting that happens
after every DDL statement, by telling the Datastax driver to not
"debounce" the events.

This reduces the total time for tests to run by about two thirds.